Process the request before creating the XHR
This means that all of the information about the request will be present (and
can therefore be taken into account) when the XHR is created. Unfortunately, it
also means that `processRequest` can no longer manipulate the XHR. If we have a
usecase for this, we can create another plugin method (i.e. `processXHR`).