Use unload and onerror where available
Though `onreadystatechange` is available in all versions of XHR, it
turns out that it doesn’t give us enough information—namely, whether
there was an error! The normal solution is to check the status (which
we were doing), but this doesn’t work for a status of 0 which can
represent an error and a success!

The new solution uses `onerror` instead of assuming that 0 statuses are
errors, when it’s available.

Closes matthewwithanm/react-inlinesvg#10